requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
gunicorn
pandas
openpyxl
click
schedule
pytz
psycopg2-binary
python-dotenv